+++ e2e/swaps/e2e-swap-suite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.12.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.4
+
 ## 1.12.3
 
 ### Patch Changes