+++ packages/pioneer/pioneer-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coinmasters/pioneer-sdk
 
+## 4.13.4
+
+### Patch Changes
+
+- bump
+
 ## 4.13.3
 
 ### Patch Changes